+++ benches/deflate.rs
@@ -3,9 +3,11 @@
 extern crate oxipng;
 extern crate test;
 
-use oxipng::internal_tests::*;
 use std::path::PathBuf;
 use test::Bencher;
+
+use oxipng::internal_tests::*;
+use oxipng::Deadline;
 
 #[bench]
 fn deflate_16_bits_strategy_0(b: &mut Bencher) {
@@ -14,7 +16,14 @@ fn deflate_16_bits_strategy_0(b: &mut Bencher) {
 
     b.iter(|| {
         let min = AtomicMin::new(None);
-        deflate(png.raw.data.as_ref(), 9, 0, 15, &min)
+        deflate(
+            png.raw.data.as_ref(),
+            9,
+            0,
+            15,
+            &min,
+            &Deadline::new(None, false),
+        )
     });
 }
 
@@ -25,7 +34,14 @@ fn deflate_8_bits_strategy_0(b: &mut Bencher) {
 
     b.iter(|| {
         let min = AtomicMin::new(None);
-        deflate(png.raw.data.as_ref(), 9, 0, 15, &min)
+        deflate(
+            png.raw.data.as_ref(),
+            9,
+            0,
+            15,
+            &min,
+            &Deadline::new(None, false),
+        )
     });
 }
 
@@ -38,7 +54,14 @@ fn deflate_4_bits_strategy_0(b: &mut Bencher) {
 
     b.iter(|| {
         let min = AtomicMin::new(None);
-        deflate(png.raw.data.as_ref(), 9, 0, 15, &min)
+        deflate(
+            png.raw.data.as_ref(),
+            9,
+            0,
+            15,
+            &min,
+            &Deadline::new(None, false),
+        )
     });
 }
 
@@ -51,7 +74,14 @@ fn deflate_2_bits_strategy_0(b: &mut Bencher) {
 
     b.iter(|| {
         let min = AtomicMin::new(None);
-        deflate(png.raw.data.as_ref(), 9, 0, 15, &min)
+        deflate(
+            png.raw.data.as_ref(),
+            9,
+            0,
+            15,
+            &min,
+            &Deadline::new(None, false),
+        )
     });
 }
 
@@ -64,7 +94,14 @@ fn deflate_1_bits_strategy_0(b: &mut Bencher) {
 
     b.iter(|| {
         let min = AtomicMin::new(None);
-        deflate(png.raw.data.as_ref(), 9, 0, 15, &min)
+        deflate(
+            png.raw.data.as_ref(),
+            9,
+            0,
+            15,
+            &min,
+            &Deadline::new(None, false),
+        )
     });
 }
 
@@ -75,7 +112,14 @@ fn deflate_16_bits_strategy_1(b: &mut Bencher) {
 
     b.iter(|| {
         let min = AtomicMin::new(None);
-        deflate(png.raw.data.as_ref(), 9, 1, 15, &min)
+        deflate(
+            png.raw.data.as_ref(),
+            9,
+            1,
+            15,
+            &min,
+            &Deadline::new(None, false),
+        )
     });
 }
 
@@ -86,7 +130,14 @@ fn deflate_8_bits_strategy_1(b: &mut Bencher) {
 
     b.iter(|| {
         let min = AtomicMin::new(None);
-        deflate(png.raw.data.as_ref(), 9, 1, 15, &min)
+        deflate(
+            png.raw.data.as_ref(),
+            9,
+            1,
+            15,
+            &min,
+            &Deadline::new(None, false),
+        )
     });
 }
 
@@ -99,7 +150,14 @@ fn deflate_4_bits_strategy_1(b: &mut Bencher) {
 
     b.iter(|| {
         let min = AtomicMin::new(None);
-        deflate(png.raw.data.as_ref(), 9, 1, 15, &min)
+        deflate(
+            png.raw.data.as_ref(),
+            9,
+            1,
+            15,
+            &min,
+            &Deadline::new(None, false),
+        )
     });
 }
 
@@ -112,7 +170,14 @@ fn deflate_2_bits_strategy_1(b: &mut Bencher) {
 
     b.iter(|| {
         let min = AtomicMin::new(None);
-        deflate(png.raw.data.as_ref(), 9, 1, 15, &min)
+        deflate(
+            png.raw.data.as_ref(),
+            9,
+            1,
+            15,
+            &min,
+            &Deadline::new(None, false),
+        )
     });
 }
 
@@ -125,7 +190,14 @@ fn deflate_1_bits_strategy_1(b: &mut Bencher) {
 
     b.iter(|| {
         let min = AtomicMin::new(None);
-        deflate(png.raw.data.as_ref(), 9, 1, 15, &min)
+        deflate(
+            png.raw.data.as_ref(),
+            9,
+            1,
+            15,
+            &min,
+            &Deadline::new(None, false),
+        )
     });
 }
 
@@ -136,7 +208,14 @@ fn deflate_16_bits_strategy_2(b: &mut Bencher) {
 
     b.iter(|| {
         let min = AtomicMin::new(None);
-        deflate(png.raw.data.as_ref(), 9, 2, 15, &min)
+        deflate(
+            png.raw.data.as_ref(),
+            9,
+            2,
+            15,
+            &min,
+            &Deadline::new(None, false),
+        )
     });
 }
 
@@ -147,7 +226,14 @@ fn deflate_8_bits_strategy_2(b: &mut Bencher) {
 
     b.iter(|| {
         let min = AtomicMin::new(None);
-        deflate(png.raw.data.as_ref(), 9, 2, 15, &min)
+        deflate(
+            png.raw.data.as_ref(),
+            9,
+            2,
+            15,
+            &min,
+            &Deadline::new(None, false),
+        )
     });
 }
 
@@ -160,7 +246,14 @@ fn deflate_4_bits_strategy_2(b: &mut Bencher) {
 
     b.iter(|| {
         let min = AtomicMin::new(None);
-        deflate(png.raw.data.as_ref(), 9, 2, 15, &min)
+        deflate(
+            png.raw.data.as_ref(),
+            9,
+            2,
+            15,
+            &min,
+            &Deadline::new(None, false),
+        )
     });
 }
 
@@ -173,7 +266,14 @@ fn deflate_2_bits_strategy_2(b: &mut Bencher) {
 
     b.iter(|| {
         let min = AtomicMin::new(None);
-        deflate(png.raw.data.as_ref(), 9, 2, 15, &min)
+        deflate(
+            png.raw.data.as_ref(),
+            9,
+            2,
+            15,
+            &min,
+            &Deadline::new(None, false),
+        )
     });
 }
 
@@ -186,7 +286,14 @@ fn deflate_1_bits_strategy_2(b: &mut Bencher) {
 
     b.iter(|| {
         let min = AtomicMin::new(None);
-        deflate(png.raw.data.as_ref(), 9, 2, 15, &min)
+        deflate(
+            png.raw.data.as_ref(),
+            9,
+            2,
+            15,
+            &min,
+            &Deadline::new(None, false),
+        )
     });
 }
 
@@ -197,7 +304,14 @@ fn deflate_16_bits_strategy_3(b: &mut Bencher) {
 
     b.iter(|| {
         let min = AtomicMin::new(None);
-        deflate(png.raw.data.as_ref(), 9, 3, 15, &min)
+        deflate(
+            png.raw.data.as_ref(),
+            9,
+            3,
+            15,
+            &min,
+            &Deadline::new(None, false),
+        )
     });
 }
 
@@ -208,7 +322,14 @@ fn deflate_8_bits_strategy_3(b: &mut Bencher) {
 
     b.iter(|| {
         let min = AtomicMin::new(None);
-        deflate(png.raw.data.as_ref(), 9, 3, 15, &min)
+        deflate(
+            png.raw.data.as_ref(),
+            9,
+            3,
+            15,
+            &min,
+            &Deadline::new(None, false),
+        )
     });
 }
 
@@ -221,7 +342,14 @@ fn deflate_4_bits_strategy_3(b: &mut Bencher) {
 
     b.iter(|| {
         let min = AtomicMin::new(None);
-        deflate(png.raw.data.as_ref(), 9, 3, 15, &min)
+        deflate(
+            png.raw.data.as_ref(),
+            9,
+            3,
+            15,
+            &min,
+            &Deadline::new(None, false),
+        )
     });
 }
 
@@ -234,7 +362,14 @@ fn deflate_2_bits_strategy_3(b: &mut Bencher) {
 
     b.iter(|| {
         let min = AtomicMin::new(None);
-        deflate(png.raw.data.as_ref(), 9, 3, 15, &min)
+        deflate(
+            png.raw.data.as_ref(),
+            9,
+            3,
+            15,
+            &min,
+            &Deadline::new(None, false),
+        )
     });
 }
 
@@ -247,7 +382,14 @@ fn deflate_1_bits_strategy_3(b: &mut Bencher) {
 
     b.iter(|| {
         let min = AtomicMin::new(None);
-        deflate(png.raw.data.as_ref(), 9, 3, 15, &min)
+        deflate(
+            png.raw.data.as_ref(),
+            9,
+            3,
+            15,
+            &min,
+            &Deadline::new(None, false),
+        )
     });
 }