+++ src/deflate/cfzlib.rs
@@ -1,9 +1,9 @@
-use Deadline;
-use atomicmin::AtomicMin;
+use crate::atomicmin::AtomicMin;
+use crate::Deadline;
+use crate::PngError;
+use crate::PngResult;
 pub use cloudflare_zlib::is_supported;
 use cloudflare_zlib::*;
-use PngError;
-use PngResult;
 
 impl From<ZError> for PngError {
     fn from(err: ZError) -> Self {
@@ -28,7 +28,7 @@ pub(crate) fn cfzlib_deflate(
     // max size is generally checked after each split,
     // so splitting the buffer into pieces gives more checks
     // = better chance of hitting it sooner.
-    let chunk_size = (data.len()/4).max(1<<15).min(1<<18); // 32-256KB
+    let chunk_size = (data.len() / 4).max(1 << 15).min(1 << 18); // 32-256KB
     for chunk in data.chunks(chunk_size) {
         stream.compress_with_limit(chunk, max_size)?;
         if deadline.passed() {
@@ -47,7 +47,8 @@ fn compress_test() {
         15,
         &AtomicMin::new(None),
         &Deadline::new(None, false),
-    ).unwrap();
-    let res = ::deflate::inflate(&vec).unwrap();
+    )
+    .unwrap();
+    let res = crate::deflate::inflate(&vec).unwrap();
     assert_eq!(&res, b"azxcvbnm");
 }