+++ src/deflate/mod.rs
@@ -1,10 +1,10 @@
-use Deadline;
-use atomicmin::AtomicMin;
-use error::PngError;
+use crate::atomicmin::AtomicMin;
+use crate::error::PngError;
+use crate::Deadline;
+use crate::PngResult;
 use miniz_oxide;
 use std::cmp::max;
 use zopfli;
-use PngResult;
 
 #[doc(hidden)]
 pub mod miniz_stream;
@@ -26,7 +26,14 @@ pub fn inflate(data: &[u8]) -> PngResult<Vec<u8>> {
 }
 
 /// Compress a data stream using the DEFLATE algorithm
-pub(crate) fn deflate(data: &[u8], zc: u8, zs: u8, zw: u8, max_size: &AtomicMin, deadline: &Deadline) -> PngResult<Vec<u8>> {
+pub(crate) fn deflate(
+    data: &[u8],
+    zc: u8,
+    zs: u8,
+    zw: u8,
+    max_size: &AtomicMin,
+    deadline: &Deadline,
+) -> PngResult<Vec<u8>> {
     #[cfg(any(target_arch = "x86_64", target_arch = "aarch64"))]
     {
         if cfzlib::is_supported() {