+++ src/filters.rs
@@ -24,35 +24,40 @@ pub fn filter_line(filter: u8, bpp: usize, data: &[u8], last_line: &[u8]) -> Vec
                 );
             };
         }
-        3 => for (i, byte) in data.iter().enumerate() {
-            if last_line.is_empty() {
-                filtered.push(match i.checked_sub(bpp) {
-                    Some(x) => byte.wrapping_sub(data[x] >> 1),
-                    None => *byte,
-                });
-            } else {
-                filtered.push(match i.checked_sub(bpp) {
-                    Some(x) => byte
-                        .wrapping_sub(((u16::from(data[x]) + u16::from(last_line[i])) >> 1) as u8),
-                    None => byte.wrapping_sub(last_line[i] >> 1),
-                });
-            };
-        },
-        4 => for (i, byte) in data.iter().enumerate() {
-            if last_line.is_empty() {
-                filtered.push(match i.checked_sub(bpp) {
-                    Some(x) => byte.wrapping_sub(data[x]),
-                    None => *byte,
-                });
-            } else {
-                filtered.push(match i.checked_sub(bpp) {
-                    Some(x) => {
-                        byte.wrapping_sub(paeth_predictor(data[x], last_line[i], last_line[x]))
-                    }
-                    None => byte.wrapping_sub(last_line[i]),
-                });
-            };
-        },
+        3 => {
+            for (i, byte) in data.iter().enumerate() {
+                if last_line.is_empty() {
+                    filtered.push(match i.checked_sub(bpp) {
+                        Some(x) => byte.wrapping_sub(data[x] >> 1),
+                        None => *byte,
+                    });
+                } else {
+                    filtered.push(match i.checked_sub(bpp) {
+                        Some(x) => byte.wrapping_sub(
+                            ((u16::from(data[x]) + u16::from(last_line[i])) >> 1) as u8,
+                        ),
+                        None => byte.wrapping_sub(last_line[i] >> 1),
+                    });
+                };
+            }
+        }
+        4 => {
+            for (i, byte) in data.iter().enumerate() {
+                if last_line.is_empty() {
+                    filtered.push(match i.checked_sub(bpp) {
+                        Some(x) => byte.wrapping_sub(data[x]),
+                        None => *byte,
+                    });
+                } else {
+                    filtered.push(match i.checked_sub(bpp) {
+                        Some(x) => {
+                            byte.wrapping_sub(paeth_predictor(data[x], last_line[i], last_line[x]))
+                        }
+                        None => byte.wrapping_sub(last_line[i]),
+                    });
+                };
+            }
+        }
         _ => unreachable!(),
     }
     filtered
@@ -64,17 +69,19 @@ pub fn unfilter_line(filter: u8, bpp: usize, data: &[u8], last_line: &[u8]) -> V
         0 => {
             unfiltered.extend_from_slice(data);
         }
-        1 => for (i, byte) in data.iter().enumerate() {
-            match i.checked_sub(bpp) {
-                Some(x) => {
-                    let b = unfiltered[x];
-                    unfiltered.push(byte.wrapping_add(b));
-                }
-                None => {
-                    unfiltered.push(*byte);
-                }
-            };
-        },
+        1 => {
+            for (i, byte) in data.iter().enumerate() {
+                match i.checked_sub(bpp) {
+                    Some(x) => {
+                        let b = unfiltered[x];
+                        unfiltered.push(byte.wrapping_add(b));
+                    }
+                    None => {
+                        unfiltered.push(*byte);
+                    }
+                };
+            }
+        }
         2 => {
             if last_line.is_empty() {
                 unfiltered.extend_from_slice(data);
@@ -113,33 +120,35 @@ pub fn unfilter_line(filter: u8, bpp: usize, data: &[u8], last_line: &[u8]) -> V
                 };
             }
         }
-        4 => for (i, byte) in data.iter().enumerate() {
-            if last_line.is_empty() {
-                match i.checked_sub(bpp) {
-                    Some(x) => {
-                        let b = unfiltered[x];
-                        unfiltered.push(byte.wrapping_add(b));
-                    }
-                    None => {
-                        unfiltered.push(*byte);
-                    }
+        4 => {
+            for (i, byte) in data.iter().enumerate() {
+                if last_line.is_empty() {
+                    match i.checked_sub(bpp) {
+                        Some(x) => {
+                            let b = unfiltered[x];
+                            unfiltered.push(byte.wrapping_add(b));
+                        }
+                        None => {
+                            unfiltered.push(*byte);
+                        }
+                    };
+                } else {
+                    match i.checked_sub(bpp) {
+                        Some(x) => {
+                            let b = unfiltered[x];
+                            unfiltered.push(byte.wrapping_add(paeth_predictor(
+                                b,
+                                last_line[i],
+                                last_line[x],
+                            )));
+                        }
+                        None => {
+                            unfiltered.push(byte.wrapping_add(last_line[i]));
+                        }
+                    };
                 };
-            } else {
-                match i.checked_sub(bpp) {
-                    Some(x) => {
-                        let b = unfiltered[x];
-                        unfiltered.push(byte.wrapping_add(paeth_predictor(
-                            b,
-                            last_line[i],
-                            last_line[x],
-                        )));
-                    }
-                    None => {
-                        unfiltered.push(byte.wrapping_add(last_line[i]));
-                    }
-                };
-            };
-        },
+            }
+        }
         _ => unreachable!(),
     }
     unfiltered