+++ src/rayon.rs
@@ -3,8 +3,13 @@ pub mod prelude {
 }
 
 pub trait ParallelIterator: Iterator + Sized {
-    fn with_max_len(self, _l: usize) -> Self {self}
-    fn reduce_with<OP>(mut self, op: OP) -> Option<Self::Item> where OP: Fn(Self::Item, Self::Item) -> Self::Item + Sync {
+    fn with_max_len(self, _l: usize) -> Self {
+        self
+    }
+    fn reduce_with<OP>(mut self, op: OP) -> Option<Self::Item>
+    where
+        OP: Fn(Self::Item, Self::Item) -> Self::Item + Sync,
+    {
         if let Some(a) = self.next() {
             Some(self.fold(a, op))
         } else {
@@ -14,18 +19,21 @@ pub trait ParallelIterator: Iterator + Sized {
 }
 
 pub trait IntoParallelIterator {
-    type Iter: Iterator<Item=Self::Item>;
+    type Iter: Iterator<Item = Self::Item>;
     type Item: Send;
     fn into_par_iter(self) -> Self::Iter;
 }
 
 pub trait IntoParallelRefIterator<'data> {
-    type Iter: Iterator<Item=Self::Item>;
+    type Iter: Iterator<Item = Self::Item>;
     type Item: Send + 'data;
     fn par_iter(&'data self) -> Self::Iter;
 }
 
-impl<I: IntoIterator> IntoParallelIterator for I where I::Item: Send {
+impl<I: IntoIterator> IntoParallelIterator for I
+where
+    I::Item: Send,
+{
     type Iter = I::IntoIter;
     type Item = I::Item;
 
@@ -35,7 +43,8 @@ impl<I: IntoIterator> IntoParallelIterator for I where I::Item: Send {
 }
 
 impl<'data, I: 'data + ?Sized> IntoParallelRefIterator<'data> for I
-    where &'data I: IntoParallelIterator
+where
+    &'data I: IntoParallelIterator,
 {
     type Iter = <&'data I as IntoParallelIterator>::Iter;
     type Item = <&'data I as IntoParallelIterator>::Item;
@@ -45,13 +54,12 @@ impl<'data, I: 'data + ?Sized> IntoParallelRefIterator<'data> for I
     }
 }
 
-impl<I: Iterator> ParallelIterator for I {
-}
+impl<I: Iterator> ParallelIterator for I {}
 
 pub fn join<A, B>(a: impl FnOnce() -> A, b: impl FnOnce() -> B) -> (A, B) {
     (a(), b())
 }
 
-pub fn spawn(a: impl FnOnce() -> A) -> A {
+pub fn spawn<A>(a: impl FnOnce() -> A) -> A {
     a()
 }